fixes for the project
