+++ Lodjinha/Lodjinha/Scenes/Home/Components/Cells/ProductTableViewCell.xib
@@ -50,7 +50,7 @@
                             <constraint firstAttribute="height" constant="1" id="7BC-Sn-Ayx"/>
                         </constraints>
                     </view>
-                    <view userInteractionEnabled="NO" contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="5bE-Zq-Jaf" customClass="SwiftDisclosureIndicator" customModule="Lodjinha" customModuleProvider="target">
+                    <view userInteractionEnabled="NO" contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="5bE-Zq-Jaf" customClass="CustomDisclosureIndicator" customModule="Lodjinha" customModuleProvider="target">
                         <rect key="frame" x="302" y="45" width="10" height="10"/>
                         <color key="backgroundColor" white="0.0" alpha="0.0" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
                         <constraints>
